+++ include/roboteam_utils/normalize.h
@@ -14,109 +14,15 @@
 
 namespace roboteam_utils {
 
-
-float mm_to_m(float scalar) {
-  return scalar/1000;
-}
-
-// rotate a ball
-void rotate(roboteam_proto::WorldBall *ball) {
-  ball->mutable_pos()->set_x(ball->pos().x()*-1);
-  ball->mutable_pos()->set_y(ball->pos().y()*-1);
-}
-
-// rotate a single robot
-void rotate(roboteam_proto::WorldRobot *robot) {
-  robot->mutable_pos()->set_x(robot->pos().x()*-1);
-  robot->mutable_pos()->set_y(robot->pos().y()*-1);
-  robot->mutable_vel()->set_x(robot->vel().x()*-1);
-  robot->mutable_vel()->set_y(robot->vel().y()*-1);
-  robot->set_w(robot->w()*-1);
-  robot->set_angle(rtt::cleanAngle(robot->angle() + M_PI));
-}
-
-// rotate the ball and robots
-void rotate(roboteam_proto::World *world) {
-  rotate(world->mutable_ball());
-
-  // rotate all blue robots
-  for (int i = 0; i < world->mutable_blue()->size(); i++) {
-    rotate(world->mutable_blue(i));
-  }
-
-  // rotate all yellow robots
-  for (int i = 0; i < world->mutable_yellow()->size(); i++) {
-    rotate(world->mutable_yellow(i));
-  }
-}
-
-// rotate the designated position given by the referee
-// this position is used for example for ball placement.
-void rotate(roboteam_proto::SSL_Referee *refereeData) {
-  refereeData->mutable_designated_position()->set_x(refereeData->designated_position().x()*-1);
-  refereeData->mutable_designated_position()->set_y(refereeData->designated_position().y()*-1);
-}
-
-// rotate a single field arc
-void rotate(roboteam_proto::SSL_FieldCicularArc *arc) {
-  arc->mutable_center()->set_x(arc->center().x()*-1);
-  arc->mutable_center()->set_y(arc->center().y()*-1);
-  arc->set_a1(arc->a1()*-1);
-  arc->set_a2(arc->a2()*-1);
-}
-
-// convert an arc from millimeters to meters
-void toMeters(roboteam_proto::SSL_FieldCicularArc *arc) {
-  arc->mutable_center()->set_x(mm_to_m(arc->center().x()));
-  arc->mutable_center()->set_y(mm_to_m(arc->center().y()));
-  arc->set_a1(mm_to_m(arc->a1()));
-  arc->set_a2(mm_to_m(arc->a2()));
-  arc->set_radius(mm_to_m(arc->radius()));
-  arc->set_thickness(mm_to_m(arc->thickness()));
-}
-
-// rotate a single field line
-void rotate(roboteam_proto::SSL_FieldLineSegment *line) {
-  line->mutable_p1()->set_x(line->p1().x()*-1);
-  line->mutable_p1()->set_y(line->p1().y()*-1);
-  line->mutable_p2()->set_x(line->p2().x()*-1);
-  line->mutable_p2()->set_y(line->p2().y()*-1);
-}
-
-// convert a line from millimeters to meters
-void toMeters(roboteam_proto::SSL_FieldLineSegment * line) {
-  line->mutable_p1()->set_x(mm_to_m(line->p1().x()));
-  line->mutable_p1()->set_y(mm_to_m(line->p1().y()));
-  line->mutable_p2()->set_x(mm_to_m(line->p2().x()));
-  line->mutable_p2()->set_y(mm_to_m(line->p2().y()));
-  line->set_thickness(mm_to_m(line->thickness()));
-}
-
-// rotate the lines and arcs of a field
-void rotate(roboteam_proto::SSL_GeometryFieldSize * field) {
-
-  // rotate all field lines
-  for (int i = 0; i < field->mutable_field_lines()->size(); i++) {
-    rotate(field->mutable_field_lines(i));
-  }
-
-  // rotate all field arcs
-  for (int i = 0; i < field->mutable_field_arcs()->size(); i++) {
-    rotate(field->mutable_field_arcs(i));
-  }
-}
-
-// convert all units from the field from millimeters to meters.
-void toMeters(roboteam_proto::SSL_GeometryFieldSize * field) {
-  // convert all field lines
-  for (int i = 0; i < field->mutable_field_lines()->size(); i++) {
-    toMeters(field->mutable_field_lines(i));
-  }
-
-  // convert all field arcs
-  for (int i = 0; i < field->mutable_field_arcs()->size(); i++) {
-    toMeters(field->mutable_field_arcs(i));
-  }
-}
+  void rotate(roboteam_proto::WorldBall *ball) ;
+  void rotate(roboteam_proto::WorldRobot *robot);
+  void rotate(roboteam_proto::World *world);
+  void rotate(roboteam_proto::SSL_Referee *refereeData);
+  void rotate(roboteam_proto::SSL_FieldCicularArc *arc);
+  void toMeters(roboteam_proto::SSL_FieldCicularArc *arc);
+  void rotate(roboteam_proto::SSL_FieldLineSegment *line);
+  void toMeters(roboteam_proto::SSL_FieldLineSegment * line);
+  void rotate(roboteam_proto::SSL_GeometryFieldSize * field);
+  void toMeters(roboteam_proto::SSL_GeometryFieldSize * field);
 
 }
